refactor: Organize directory structure
- Make cmd entrypoint minimal
- Create public and private parts of code
