+++ pkg/utils/tools.go
@@ -1,0 +1,18 @@
+package utils
+
+import (
+	"fmt"
+	"os/exec"
+)
+
+func ValidateTools(tools []string) bool {
+	allAvailable := true
+	for _, tool := range tools {
+		_, err := exec.LookPath(tool)
+		if err != nil {
+			allAvailable = false
+			fmt.Printf("Failed finding %s package: %s", tool, err)
+		}
+	}
+	return allAvailable
+}
